fasfapi_cli/tpl/project/{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = []
license = "Apache-2.0"


[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
watchgod = "^0.6"
uvicorn = "^0.13"
python-multipart = "^0.0.5"
python-jose = "^3.2.0"
passlib = "^1.7.4"
aiofiles = "^0.6.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "^20.8b1"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# 谨慎使用，会导致包更新不及时
[[tool.poetry.source]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
name = "tuna"
default = true
